# pkg/api/handlers/openapi.yaml
openapi: "3.0.3"
info:
  title: driftwire API
  description: |
    Real-time, multi-cloud Terraform drift detection API with Falco integration.
    Supports AWS CloudTrail, GCP Audit Logs, and Azure Activity Logs.
  version: "0.9.0"
  contact:
    name: Keita Higaki
    url: https://github.com/higakikeita/driftwire
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/state/resources:
    get:
      tags: [State]
      summary: List Terraform resources
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Paginated list of Terraform resources
...